+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/lost_found_board/LostFoundBoardImage.java
@@ -6,10 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 
 @Entity
@@ -21,6 +18,7 @@ public class LostFoundBoardImage {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(nullable = false)
     private Long lostFoundBoardId;
 
     private ImageFile imageFile;
