+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/web/api/PageApiController.java
@@ -3,17 +3,23 @@ package kr.ac.kumoh.illdang100.tovalley.web.api;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.ac.kumoh.illdang100.tovalley.dto.ResponseDto;
+import kr.ac.kumoh.illdang100.tovalley.dto.lost_found_board.LostFoundBoardReqDto;
+import kr.ac.kumoh.illdang100.tovalley.dto.lost_found_board.LostFoundBoardRespDto;
 import kr.ac.kumoh.illdang100.tovalley.security.auth.PrincipalDetails;
 import kr.ac.kumoh.illdang100.tovalley.service.page.PageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
 
 import static kr.ac.kumoh.illdang100.tovalley.dto.page.PageRespDto.*;
 
@@ -54,5 +60,24 @@ public class PageApiController {
         MyPageAllRespDto result = pageService.getMyPageAllData(memberId, pageable);
 
         return new ResponseEntity<>(new ResponseDto<>(1, "마이 페이지 조회에 성공하였습니다", result), HttpStatus.OK);
+    }
+
+    @GetMapping("/lostItem")
+    public ResponseEntity<?> getLostFoundBoardList(@ModelAttribute @Valid LostFoundBoardReqDto.LostFoundBoardListReqDto lostFoundBoardListReqDto,
+                                                   BindingResult bindingResult,
+                                                   @PageableDefault(size = 5, sort = "createdDate", direction = Sort.Direction.DESC) Pageable pageable) {
+
+        Slice<LostFoundBoardRespDto.LostFoundBoardListRespDto> lostFoundBoardList = pageService.getLostFoundBoardList(lostFoundBoardListReqDto, pageable);
+
+        return new ResponseEntity<>(new ResponseDto<>(1, "분실물 찾기 페이지 조회를 성공했습니다", lostFoundBoardList), HttpStatus.OK);
+    }
+
+    @GetMapping("/auth/lostItem/{lostFoundBoardId}")
+    public ResponseEntity<?> getLostFoundBoardDetail(@PathVariable long lostFoundBoardId,
+                                                     @AuthenticationPrincipal PrincipalDetails principalDetails) {
+
+        LostFoundBoardRespDto.LostFoundBoardDetailRespDto lostFoundBoardDetail = pageService.getLostFoundBoardDetail(lostFoundBoardId, principalDetails.getMember().getEmail());
+
+        return new ResponseEntity<>(new ResponseDto<>(1, "분실물 찾기 상세 페이지 조회를 성공했습니다", lostFoundBoardDetail), HttpStatus.OK);
     }
 }
